+++ src/main/scala/tscfg/javaGenerator.scala
@@ -39,7 +39,7 @@ object javaGenerator {
       }
 
       def genForBranch(bn: BranchNode): Unit = {
-        val className = upperFirst(symbol)
+        val className = getClassName(symbol)
 
         if (!isRoot) {
           // declare symbol:
@@ -70,7 +70,7 @@ object javaGenerator {
               out.println(s"""$instance;""")
 
             case BranchNode(k, _)  =>
-              val className = upperFirst(k.simple)
+              val className = getClassName(k.simple)
               out.println(s"""new $className(__$$config(c, "${k.simple}"));""")
           }
         }
@@ -155,20 +155,28 @@ object javaGenerator {
     * Returns a valid Java identifier from the given symbol:
     *
     * - appends a '_' in case the symbol is a java keyword or special literal ("null", "true", "false");
+    * - otherwise, prepends a '_' in case the symbol is numerical literal;
     * - otherwise, returns the given symbol if already a valid java identifier;
     * - otherwise, prefixes the symbol with '_' if first character is valid but not at first position, and
     *   replaces any character that cannot be part of a java identifier with '_'.
     */
   def javaIdentifier(symbol: String): String = {
-    if (javaKeywords.contains(symbol)) symbol + "_"
-    else if (isJavaIdentifier(symbol)) symbol else {
+    if (javaKeywords.contains(symbol))
+      symbol + "_"
+    else if (isJavaIdentifier(symbol))
+      symbol
+    else {
       val c0 = symbol.charAt(0)
-      val first: String = if (Character.isJavaIdentifierStart(c0)) String.valueOf(c0)
-      else if (Character.isJavaIdentifierPart(c0)) "_" + c0 else "_"
-      val rest = symbol.substring(1) map { c =>
-        if (Character.isJavaIdentifierPart(c)) c else '_'
+      if (Character.isDigit(c0))
+        "_" + symbol
+      else {
+        val first: String = if (Character.isJavaIdentifierStart(c0)) String.valueOf(c0)
+        else if (Character.isJavaIdentifierPart(c0)) "_" + c0 else "_"
+        val rest = symbol.substring(1) map { c =>
+          if (Character.isJavaIdentifierPart(c)) c else '_'
+        }
+        first + rest
       }
-      first + rest
     }
   }
 
@@ -176,6 +184,8 @@ object javaGenerator {
     Character.isJavaIdentifierStart(symbol.charAt(0)) &&
       symbol.substring(1).forall(Character.isJavaIdentifierPart)
   }
+
+  private def getClassName(symbol:String) = upperFirst(javaIdentifier(symbol))
 
   private def upperFirst(symbol:String) = symbol.charAt(0).toUpper + symbol.substring(1)